# Optimize String#to_u methods for the case of a negative number