This makes `String`'s `to_u64_info` method return early when the number that should be converted to an unsigned integer is negative.
```crystal
Benchmark.ips do |bm|
  bm.report("old") do
    "-112978314788778".to_u64 { false }
  end

  bm.report("new") do
    "-112978314788778".new_to_u64 { false }
  end
end
```
```
old  58.13M (  17.2ns) (± 5.40%)  0 B/op  10.21× slower
new  593.7M (  1.68ns) (± 4.43%)  0 B/op        fastest
```

---

```crystal
Benchmark.ips do |bm|
  bm.report("old") do
    "-112978314788778913431478989314731437891343734789943413789474187178137914713173484137134780".to_u16 { false }
  end

  bm.report("new") do
    "-112978314788778913431478989314731437891343734789943413789474187178137914713173484137134780".new_to_u16 { false }
  end
end
```
```
old  43.73M ( 22.87ns) (± 8.58%)  0 B/op  13.42× slower
new 586.69M (   1.7ns) (± 3.40%)  0 B/op        fastest
```