@RX14 2019-05-21
It improves performance in an edge case only, but that doesn't mean it's not worth merging, as it's only a small bit of code.